requirements.txt
colorama
netifaces
requests
PyYAML
marshmallow
dotty-dict
deepmerge